Add files via upload
The beginning of big changes!
Instead of simply substituting WORDS for instructions. Here the word stores the TOKEN in the token queue. There it will be processed for possible combos. And only then the token queue will gradually be replaced by instructions, as it was done before.
The potential is great. Because in the future the word will guide what was before it and what will be after it. If we used an index in the loop or not, etc.